Data.Expression.Z3

Documentation

class IFunctor f => IToZ3 f

Minimal complete definition

itoZ3

toZ3 :: forall f (s :: Sort) z3. (IToZ3 f, MonadZ3 z3) => IFix f s -> z3 AST

class IFromZ3Into f f => IFromZ3 f

fromZ3 :: forall (f :: (Sort -> *) -> Sort -> *) (s :: Sort) z3. (IFromZ3 f, MonadZ3 z3, SingI s) => AST -> z3 (IFix f s)

assert :: forall (f :: (Sort -> *) -> Sort -> *) z3. (IToZ3 f, MonadZ3 z3) => IFix f BooleanSort -> z3 ()

model :: forall (f :: (Sort -> *) -> Sort -> *) (s :: Sort) z3. (IToZ3 f, IFromZ3 f, IShow f, MonadZ3 z3, SingI s) => IFix f s -> z3 (IFix f s)

unsatcore :: forall (f :: (Sort -> *) -> Sort -> *) z3. (IToZ3 f, MonadZ3 z3) => [IFix f BooleanSort] -> z3 [IFix f BooleanSort]

interpolate :: forall (f :: (Sort -> *) -> Sort -> *) z3. (IToZ3 f, IFromZ3 f, MonadZ3 z3) => [IFix f BooleanSort] -> z3 [IFix f BooleanSort]

eliminate :: forall (f :: (Sort -> *) -> Sort -> *) z3. (IToZ3 f, IFromZ3 f, MonadZ3 z3) => IFix f BooleanSort -> z3 (IFix f BooleanSort)

local :: MonadZ3 z3 => z3 a -> z3 a

Run a query and restore the initial logical context.

This is a shorthand for push, run the query, and pop.

push :: MonadZ3 z3 => z3 ()

Create a backtracking point.

For push; m; pop 1 see local.

pop :: MonadZ3 z3 => Int -> z3 ()

Backtrack n backtracking points.

Contrary to solverPop this funtion checks whether n is within the size of the solver scope stack.

check :: MonadZ3 z3 => z3 Result

Check whether the given logical context is consistent or not.

data Result

Result of a satisfiability check.

This corresponds to the z3_lbool type in the C API.

Constructors

Sat

Unsat

Undef